Add abstract BaseSubworkflowNode class
